Refactor Joystick to individual module.
